Adds GUID to manifest records for tracking
Adds a GUID to the processed_manifests table and assigns a unique GUID to each manifest record to enhance tracking and identification.

This change also includes refactoring to use the GUID in log messages and database operations, which improves data integrity and debugging capabilities.

Additionally, adds ez-guid package as a dependency, removes unused schema.sql, and updates .gitignore.

Fixes a bug where the acknowledgement email was always reporting the vendor as Azura Fresh even when it wasn't